refactor: move expressions package
