+++ exact-numbers/src/main/kotlin/xyz/lbres/expressions/term/ref.kt
@@ -1,0 +1,5 @@
+// ktlint-disable filename
+package xyz.lbres.expressions.term
+
+@Deprecated("Class moved in v1.0.0", ReplaceWith("Term", "xyz.lbres.exactnumbers.expressions.term.Term"), DeprecationLevel.ERROR)
+typealias Term = xyz.lbres.exactnumbers.expressions.term.Term
