+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/expressions/term/binaryOpsTests.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.expressions.term
+package xyz.lbres.exactnumbers.expressions.term
 
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
 import xyz.lbres.exactnumbers.irrationals.log.Log
